Ajout de la confirmation de suppression de saison

+++ GestAssoc/GestAssoc.Modules.GestionSaisons/Commands/DeleteSaisonCommand.cs
@@ -4,6 +4,7 @@ using GestAssoc.Model.Models;
 using GestAssoc.Modules.GestionSaisons.Constantes;
 using GestAssoc.Modules.GestionSaisons.Properties;
 using GestAssoc.Modules.GestionSaisons.Services;
+using Microsoft.Practices.Prism.Interactivity.InteractionRequest;
 using Microsoft.Practices.ServiceLocation;
 using Microsoft.Practices.Unity;
 using System;
@@ -15,6 +16,14 @@ namespace GestAssoc.Modules.GestionSaisons.Commands
 {
 	public class DeleteSaisonCommand : ICommand
 	{
+		public InteractionRequest<IConfirmation> RqConfirmDelete { get; private set; }
+		private Action _commandCallBack;
+
+		public DeleteSaisonCommand(Action callback) {
+			this.RqConfirmDelete = new InteractionRequest<IConfirmation>();
+			this._commandCallBack = callback;
+		}
+		
 		public bool CanExecute(object parameter) {
 			var itemToDelete = parameter as Saison;
 
@@ -29,35 +38,33 @@ namespace GestAssoc.Modules.GestionSaisons.Commands
 		}
 
 		public void Execute(object parameter) {
-			var service = ServiceLocator
-				.Current.GetInstance<IUnityContainer>()
-				.Resolve<IGestionSaisonsServices>();
+			this.RqConfirmDelete.Raise(
+				new Confirmation 
+				{ 
+					Content = GlblRes.Confirm_SuppressionSaison + Environment.NewLine + (parameter as Saison).ToString(), 
+					Title = Common.Properties.Resources.Titre_Confirmation
+				},
+				c => this.ExecuteCallback(c.Confirmed, parameter as Saison)
+			);
+		}
 
-			var itemToDelete = parameter as Saison;
+		private void ExecuteCallback(bool deleteConfirmed, Saison itemToDelete) {
+			if (deleteConfirmed) {
+				var service = ServiceLocator
+					.Current.GetInstance<IUnityContainer>()
+					.Resolve<IGestionSaisonsServices>();
 
-			try {
-				UIServices.SetBusyState();
-				service.DeleteSaison(itemToDelete);
-
-				NotificationHelper.WriteLog(GlblRes.Log_ElementSupprime);
-
-				// La saison la plus récente devient la saison courante
-				UIServices.SetBusyState();
-				var newSaisonCourante = service.GetAllSaisons().LastOrDefault();
-
-				if (newSaisonCourante != null) {
+				try {
 					UIServices.SetBusyState();
-					service.SetSaisonCourante(newSaisonCourante);
-					NotificationHelper.WriteLog(Resources.Log_NouvelleSaisonCourante + newSaisonCourante.ToString());
-				}
-				else {
-					NotificationHelper.WriteLog(GlblRes.Log_AucuneSaisonCourante);
-				}
+					service.DeleteSaison(itemToDelete);
 
-				new ShowViewCommand(ViewNames.ConsultationSaisons.ToString()).Execute(null);
-			}
-			catch (Exception ex) {
-				NotificationHelper.ShowError(ex);
+					NotificationHelper.WriteLog(GlblRes.Log_EnregistrementSupprime);
+
+					this._commandCallBack();
+				}
+				catch (Exception ex) {
+					NotificationHelper.ShowError(ex);
+				}
 			}
 		}
 	}
